rm exporouter and add react router
