--- frontend/app/(auth)/_layout.tsx
@@ -1,14 +1,0 @@
-import { Stack } from 'expo-router';
-
-
-const Layout = () => {
-  return (
-      <Stack>
-        <Stack.Screen name="welcome" options={{ headerShown: false }} />
-        <Stack.Screen name="sign-up" options={{ headerShown: false }} />
-        <Stack.Screen name="sign-in" options={{ headerShown: false }} />
-      </Stack>
-  );
-}
-
-export default Layout;
