+++ frontend/tsconfig.json
@@ -12,14 +12,12 @@
     "**/*.ts",
     "**/*.tsx",
     "./app/**/*.tsx",
-    ".expo/types/**/*.ts",
-    "expo-env.d.ts",
     "src/**/*.ts",
     "src/**/*.tsx",
     "index.tsx",
     "/App/**/*.ts",
     "App.tsx",
     "app.d.ts",
-    "App/**/*.tsx",
+    "App/**/*.tsx"
   ]
 }
